Add WishlistEntries property to Product class
Introduced a new `WishlistEntries` property to the `Product` class in `Product.cs`. This property is of type `ICollection<Wishlist>` and is initialized to a new `List<Wishlist>`. It includes XML documentation describing it as a collection of wishlist entries for the product.

This change extends the `Product` class functionality to support managing wishlist entries, which can be useful in scenarios like e-commerce applications. No existing functionality was modified or removed.